Configures image directory and updates dependencies
Configures the image directory using an environment variable to allow for flexible image storage locations.

Updates the docker-compose file to mount the image directory into the container.

Additionally, updates the image processing command to use the executable directly, assuming it is on the PATH.
Also adds images/ folder to .gitignore.