Improves session list item interaction
Changes the session list item to use a div element, instead of an anchor element, which allows the image thumbnail to also act as a link to the session details.

This provides a better user experience by making the entire session item clickable, and also fixes a bug where the thumbnail image was not clickable.